pyproject.toml
# To check for missing dependencies, use pipreqs:
# pipx install pipreqs  # one-time installation
# pipreqs . --print    # shows all imports without creating a file
# Then add any missing packages with: uv add package-name

[project]
name = "dsg"
version = "0.4.2"
description = "Data Sync Gizmo - HRDAG's data versioning system"
authors = [
    { name = "Sam Z", email = "shimian.zhang@gmail.com" },
    { name = "Patrick Ball", email = "pball@hrdag.org" }
]
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Private :: Do Not Upload"
]
dependencies = [
    "typer>=0.15.3,<0.16.0",
    "tomlkit>=0.13.2,<0.14.0",
    "loguru>=0.7.3,<0.8.0",
    "xxhash>=3.5.0,<4.0.0",
    "orjson>=3.9.15,<4.0.0",
    "pydantic>=2.0,<3.0",
    "rich>=13.0,<14.0",
    "pyyaml>=6.0,<7.0",
    "email-validator>=2.0,<3.0",
    "lz4>=4.3.2,<5.0.0",
    "paramiko>=3.0,<4.0",
    "humanize>=4.0,<5.0",
    # AI tools dependencies
    "PyGithub>=1.59.0,<2.0.0",
    "click>=8.1.0,<9.0.0",
    "jinja2>=3.1.0,<4.0.0",
]

[project.scripts]
dsg = "dsg.cli:main"
dsg-ai-tools = "dsg_ai_tools.cli:main"

[project.urls]
Homepage = "https://github.com/hrdag/dsg"
Repository = "https://github.com/hrdag/dsg"
"Bug Tracker" = "https://github.com/hrdag/dsg/issues"

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=8.3.5",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "expects>=0.9.0",
    "pyfakefs>=5.8.0",
    "ruff>=0.12.0",
]

[tool.uv.sources]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dsg"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

